Riki/requirements.txt
alembic==1.12.0
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
dominate==2.8.0
exceptiongroup==1.1.3
Faker==12.0.1
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mixer==7.2.2
outcome==1.2.0
packaging==23.1
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==8.0.1
pytz==2023.3.post1
sdist==0.0.0
selenium==4.12.0
six==1.16.0
slugify==0.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.21
text-unidecode==1.3
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
urllib3==2.0.5
visitor==0.1.3
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
xvfbwrapper==0.2.9
pdfdocument==4.0.0
markdown2==2.4.10
python-docx==1.1.0
email-validator==1.1.3
